+++ vm/include/RegisterManager.hpp
@@ -9,22 +9,23 @@
 #define KOTO_VM_REGISTER_MANAGER_H_DEFINED
 
 #include <cstdlib>
+#include <cstdint>
 
 class RegisterManager {
     public:
-        unsigned int& operator[](size_t index);
+        uint16_t& operator[](size_t index);
 
         bool getBit(size_t index, size_t bit);
         void setBit(size_t index, size_t bit, bool value);
         void togBit(size_t index, size_t bit);
 
-        unsigned int& arith1 = registers[11];
-        unsigned int& arith2 = registers[12];
-        unsigned int& flags  = registers[13];
-        unsigned int& stack  = registers[14];
-        unsigned int& pc     = registers[15];
+        uint16_t& arith1 = registers[11];
+        uint16_t& arith2 = registers[12];
+        uint16_t& flags  = registers[13];
+        uint16_t& stack  = registers[14];
+        uint16_t& pc     = registers[15];
     private:
-        unsigned int registers[16];
+        uint16_t registers[16];
 };
 
 #endif
