+++ vm/source/MemoryManager.cpp
@@ -1,11 +1,12 @@
-#include <MemoryManager.hpp>
+#include "MemoryManager.hpp"
 
 #include <cstdlib>
+#include <cstdint>
 #include <cstring>
 #include <stdexcept>
 
 MemoryManager::MemoryManager(std::size_t words) {
-    this->data = (unsigned int*) malloc(sizeof(unsigned int) * words);
+    this->data = (uint16_t*) malloc(sizeof(uint16_t) * words);
     this->words = words;
 }
 
@@ -13,15 +14,15 @@ MemoryManager::~MemoryManager() {
     free(this->data);
 }
 
-void MemoryManager::setRange(std::size_t index, unsigned int* values, std::size_t length) {
+void MemoryManager::setRange(std::size_t index, uint16_t* values, std::size_t length) {
     if (index + length >= this->words) {
         throw std::out_of_range("Memory::setRange");
     }
 
-    memcpy(this->data + index, values, sizeof(unsigned int) * length);
+    memcpy(this->data + index, values, sizeof(uint16_t) * length);
 }
 
-unsigned int& MemoryManager::operator[](std::size_t index) {
+uint16_t& MemoryManager::operator[](std::size_t index) {
     if (index >= this->words) {
         throw std::out_of_range("Memory::operator[]");
     }
